feat(theming): fixing PR review comments
