+++ superset-frontend/src/embedded/index.tsx
@@ -21,7 +21,12 @@ import 'src/public-path';
 import { lazy, Suspense } from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
-import { makeApi, t, logging } from '@superset-ui/core';
+import {
+  type SupersetThemeConfig,
+  makeApi,
+  t,
+  logging,
+} from '@superset-ui/core';
 import Switchboard from '@superset-ui/switchboard';
 import getBootstrapData, { applicationRoot } from 'src/utils/getBootstrapData';
 import setupClient from 'src/setup/setupClient';
@@ -37,7 +42,6 @@ import {
   EmbeddedContextProviders,
   getThemeController,
 } from './EmbeddedContextProviders';
-import { SupersetThemeConfig } from '../theme/ThemeController';
 import { embeddedApi } from './api';
 import { getDataMaskChangeTrigger } from './utils';
